Optimized the code
On line 19:
You need not compare the return value with 0. If Serial.available() returns any value greater than 0, control enters the while loop. Any value greater than 0 is treated as 1.

On lines 11, 20 & 24:
You can directly pass the return value of Serial.read() by calling the method inside the switch itself (line 20).
If this approach is used, then it is unnecessary to create and initialize the 'command' variable.

I have made the above changes and commented the unnecessary lines of code.